refactor environment target types to their own module
